requirements.txt
qrcode
django
pillow
pytz
XlsxWriter
django-case-insensitive-field
django-settings-export
djangorestframework
